% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaklist_browser.R
\name{add_view_col}
\alias{add_view_col}
\alias{shinyInput}
\title{Function to add the View button}
\usage{
add_view_col(X, offset, modal_col)

shinyInput(FUN, len, id, offset, label, ...)
}
\arguments{
\item{X}{The table to add buttons to.}

\item{offset}{A numeric vector. Offset for the button id for each row.}

\item{modal_col}{A string. The second nesting for feature annotations.}

\item{FUN}{The shiny Input function to create.}

\item{len}{Integer How many buttons to make.}

\item{id}{String. Prefix for the id.}

\item{label}{A character vector. The text to but on each button. Should have the same length as the number of rows in X.}

\item{...}{}
}
\value{
Callback stript.
}
\description{
Internal function for the datatables View button that opens the modal
}
\section{Functions}{
\itemize{
\item \code{add_view_col}: Generates all buttons

\item \code{shinyInput}: generates multople shiny inputs with different labels
}}

\keyword{internal}
