% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaklist_browser.R
\name{BP_callback}
\alias{BP_callback}
\title{Datatables callback function to modify peakbrowser table}
\usage{
BP_callback(modal_col)
}
\arguments{
\item{modal_col}{A string. The second nesting for feature annotations.}
}
\value{
Callback stript.
}
\description{
Internal function for the database callback
}
\keyword{internal}
