% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaklist_annotation.R
\name{cmp_mz_filter}
\alias{cmp_mz_filter}
\title{Compound list filter}
\usage{
cmp_mz_filter(mz, ref_tbl, mz_col = "mz", ppm = 30)
}
\arguments{
\item{mz}{Query mz to filter by}

\item{ref_tbl}{A \code{\link{tibble}} with compound mz values. Can be generated with expand_adducts.}

\item{mz_col}{Character string. Column name of the column containing the mz values.}

\item{ppm}{Maximum ppm value between query mass and database mz.}
}
\value{
A \code{\link{tibble}} containing the same columns as the input table but having add a ppm column for the deviation between compound mz and query mass.
}
\description{
This function filters a compound list based on mz value.
It is meant to be used to annotate a peaklist by running through (with \code{\link{map}}) 
each mz in the peaklist and adding a nested table of hits to the peaklist.
}
