% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{cmp_tbl_pretty}
\alias{cmp_tbl_pretty}
\title{Prepare compound table for interactive display}
\usage{
cmp_tbl_pretty(cmp_tbl)
}
\arguments{
\item{cmp_tbl}{\code{\link{tibble}} of compounds.}
}
\value{
A \code{\link{tibble}} containing the same columns as the input table.
}
\description{
This function modifies some of the columns in a compound table to be displayed better in an interactive table.
To the inchi column code to crop very long strings will be added.
To the name column "; " separater will be changed to a line break.
This is mainly used internally.
}
