% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{do.aromaticity_p}
\alias{do.aromaticity_p}
\title{Robust helper functions to process structures with rcdk}
\usage{
\method{do}{aromaticity_p}(...)
}
\value{
Returns functions equavalent of \code{\link{do.aromaticity}}, \code{\link{do.typing}}, \code{\link{do.isotopes}}, \code{\link{get.exact.mass}} but made robust with \code{\link{possibly}}.
}
\description{
Some helper functions modified from rcdk to return NA on failure.
This makes it easier to process many compounds where some might behave unexpected.
}
\keyword{internal}
