% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_adducts.R
\name{expand_adducts}
\alias{expand_adducts}
\title{Expand mass to list of adducts}
\usage{
expand_adducts(cmp_tbl, mode = "pos", adducts = c("[M+H]+", "[M+Na]+",
  "[2M+H]+", "[M+K]+", "[M+H-H2O]+"))
}
\arguments{
\item{cmp_tbl}{data.frame of compounds. The tables should contain a column named "mass".}

\item{mode}{A string. Either "pos", "neg" or "ei".}

\item{adducts}{Character vector. Any adduct listed in the adduct list found here: https://github.com/stanstrup/chemhelper/tree/master/inst/extdata. Examples: "[M+H]+", "[M+Na]+", "[M-H]-", "[M+Cl]-".}
}
\value{
A \code{\link{tibble}} containing the same columns as the input table but with added columns: adduct, charge, nmol, mz, mode.
}
\description{
Expand a table with masses to a table with selected adducts and fragments and their mz values.
This function takes a table with a column mass and calculates the mz value of all selected adducts and fragments.
}
