% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{generate_lipidblast_tbl}
\alias{generate_lipidblast_tbl}
\title{Generate table with LipidBlast compounds}
\usage{
generate_lipidblast_tbl(json_path)
}
\arguments{
\item{json_path}{Path to the LipidBlast json file available at: http://mona.fiehnlab.ucdavis.edu/downloads}
}
\value{
A \code{\link{tibble}} containing the columns: id, name, inchi, formula, and mass.
}
\description{
Generate table with LipidBlast compounds
}
\examples{
\dontrun{
lipidblast_tbl <- generate_lipidblast_tbl("path/to/MoNA-export-LipidBlast.json")

# Alternatively get pre-made table from the package
lipidblast_tbl2 <- readRDS(system.file("extdata", "lipidblast_tbl.rds", package="PeakABro"))
}


}
