% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{generate_lipidmaps_tbl}
\alias{generate_lipidmaps_tbl}
\title{Generate table with LipidMaps compounds}
\usage{
generate_lipidmaps_tbl(sdf_path)
}
\arguments{
\item{sdf_path}{Path to the Lipidmaps SDF file available at: http://www.lipidmaps.org/resources/downloads/}
}
\value{
A \code{\link{tibble}} containing the columns: id, name, inchi, formula, and mass.
}
\description{
Generate table with LipidMaps compounds
}
\examples{
\dontrun{
lipidmaps_tbl <- generate_lipidmaps_tbl("path/to/LMSDFDownload6Dec16FinalAll.sdf")

# Alternatively get pre-made table from the package
lipidmaps_tbl2 <- readRDS(system.file("extdata", "lipidmaps_tbl.rds", package="PeakABro"))
}


}
