% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{inchi2formula}
\alias{inchi2formula}
\title{Inchi to formula}
\usage{
inchi2formula(inchi)
}
\arguments{
\item{inchi}{Character vector with inchis}
}
\value{
Character vector
}
\description{
Inchi to formula
}
