% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaklist_annotation.R
\name{.is_between_1range}
\alias{.is_between_1range}
\alias{.is_between}
\title{is_between}
\usage{
.is_between_1range(x, a, b)

.is_between(x, a, b)
}
\arguments{
\item{x}{numeric vector to check.}

\item{a}{lower limit of interval. Scalar for is_between_1range. Can be a vector for is_between.}

\item{b}{upper limit of interval. Scalar for is_between_1range. Can be a vector for is_between.}
}
\value{
Logical vector (is_between_1range) of length x or 
logical matrix (is_between) of dimensions x times length of a (== length of b).
}
\description{
is_between
}
