% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{LB_getname}
\alias{LB_getname}
\alias{LB_get_inchi}
\alias{LB_get_id}
\title{Helper functions to extract info from LipidBlast json.}
\usage{
LB_getname(json)

LB_get_inchi(json)

LB_get_id(json)
}
\arguments{
\item{json}{LipidBlast json}
}
\value{
Returns character vector with the relevant info.
}
\description{
Some helper functions to extract info from LipidBlast json.
}
\section{Functions}{
\itemize{
\item \code{LB_getname}: Extracts the compound name from LipidBlast json

\item \code{LB_get_inchi}: Extracts the inchi from LipidBlast json

\item \code{LB_get_id}: Extracts the id name from LipidBlast json
}}

\keyword{internal}
