% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaklist_browser.R
\name{peaklist_browser}
\alias{peaklist_browser}
\title{Peaklist Browser}
\usage{
peaklist_browser(peaklist, collapse_col, modal_col)
}
\arguments{
\item{peaklist}{A \code{\link{tibble}} peaklist with nested columns for pcgroup and the feauture annotations}

\item{collapse_col}{A string. The first nesting for pcgroups.}

\item{modal_col}{A string. The second nesting for feature annotations.}
}
\value{
Opens a shiny app with the peaklist browser
}
\description{
Open a peaklist browser for a peaklist prepared with \code{\link{peaklist_browser_prep}}.
}
\references{
https://github.com/rstudio/shiny-examples/issues/9#issuecomment-295018270

https://stackoverflow.com/questions/38575655/shiny-modules-namespace-outside-of-ui-for-javascript-links

https://stackoverflow.com/questions/45739303/r-shiny-handle-action-buttons-in-data-table

https://stackoverflow.com/questions/40107022/r-shinybs-popup-window

https://stackoverflow.com/questions/46593607/datatable-with-nesting-child-rows-and-modal
}
