% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{do.aromaticity_p}
\alias{do.aromaticity_p}
\alias{do.typing_p}
\alias{do.isotopes_p}
\alias{get.exact.mass_p}
\alias{parse.inchi_p}
\title{Robust helper functions to process structures with rcdk}
\usage{
do.aromaticity_p(...)

do.typing_p(...)

do.isotopes_p(...)

get.exact.mass_p(...)

parse.inchi_p(...)
}
\value{
Returns functions equavalent of \code{\link{do.aromaticity}}, \code{\link{do.typing}}, \code{\link{do.isotopes}}, \code{\link{get.exact.mass}} but made robust with \code{\link{possibly}}.
}
\description{
Some helper functions modified from rcdk to return NA on failure.
This makes it easier to process many compounds where some might behave unexpected.
}
\section{Functions}{
\itemize{
\item \code{do.aromaticity_p}: Robust version of \code{\link{do.aromaticity}}

\item \code{do.typing_p}: Robust version of \code{\link{do.typing}}

\item \code{do.isotopes_p}: Robust version of \code{\link{do.isotopes}}

\item \code{get.exact.mass_p}: Robust version of \code{\link{get.exact.mass}}

\item \code{parse.inchi_p}: Robust version of \code{\link{parse.inchi}}
}}

\keyword{internal}
